Stop generating old synthetics
